tools: add the c compiler stuff
There is a long way to go on this. It can compile some integer and long stuff
but there is a lot left to fill in, from long compares to teaching the
assembler how to do self extending branches properly and some kind of
support for auto-switching from PC relative jump to long ones.