cc: bit more work
- Realign the instruction names to the new ones
- Some work on register pressure and store

It's still not enough. Y causes us all kinds of problems. Might be worth
an experiment to see how much worse it is if we reserve Y for 32bit.

If not we are going to need some ugly helpers to implement the equivalent of
st y,(-s) and maybe clobber A or something to make space for the helper ?